requirements.txt
spacy==2.0.16
sklearn
pytesseract
pdf2image
pillow
poppler
